requirements.txt
celery==4.0.2
Django==1.10.7
django-extensions==1.7.9
django_celery_results
django-filter
djangorestframework==3.6.2
djangorestframework-csv==2.0.0
flower
gunicorn==19.7.1
psycopg2==2.7.1
sutime
tweepy


